AudioScope. Bash audio oscilloscope script.
This is an audio oscilloscope using the sound card in bash scripting
only.
It is purely text mode and the only dependency for FULL operation is an
installation of SoX, minimum version 14.4.0.
It defaults to DEMO mode on first start and in this mode is designed to
work on almost any *NIX and Linux machine available; /dev/dsp is also
included for lesser machines.
It is designed around a Macbook Pro 13" OSX 10.7.5 Circa August 2012
using the default bash terminal and tested also on PCLinuxOS 2009 and
Debian 6.0.x.
There is lots of information in the code itself including simple
circuits so make sure to read the code first…