Added DC input using Arduino at the moment and others.
Lots of changes here for version 0.22.10...

A semi-automatic Arduino detection command:- ARDDET<CR>
This requires a small amount of user input.
The ARDDET<CR> section also creates a .PDE file to program the Arduino
for a single ADC only <- read the code for more info...
To see Arduino working after programming it just temporarily wire a 10K
pot' between +5V and Gnd with the wiper connected to ANALOG 0.

An 8 second pure sinewave wave file now created for multi-platform use.
It IS now on my 'phone as a signal source using the default 'phone
player in loop mode - nice'n'easy... ;o)

Another command:- BURST<CR> plays the above wave file using "afplay"
for OSX or "aplay" for Linux.

DC now allows REAL DC into the scope IF Arduino is detected.
IF NOT then it defaults to a DEMO mode and samples /dev/urandom.
It displays itself inside the YELLOW section of the COMMAND window for
a short period.

Another command:- ALTDC<CR> is a placeholder for a sound card DC input
under development.

Other minor changes too...